+++ src/filename_isdir.cxx
@@ -54,7 +54,7 @@ int fl_filename_isdir(const char* n) {
   char		fn[FL_PATH_MAX];
   int		length;
 
-  length = strlen(n);
+  length = (int) strlen(n);
 
 #ifdef WIN32
   // This workaround brought to you by the fine folks at Microsoft!
